requirements.txt
gradio>=4.0.0
stability-sdk>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
protobuf>=4.0.0 